clients/cli/requirements.txt
aiohttp==3.13.2
PyNaCl==1.5.0
